put utils of ordinary  mnl bandit in file ordinary_mnl_bandit_utils.py
