+++ banditpylib/bandits/__init__.py
@@ -1,6 +1,7 @@
 from .utils import *
 from .ordinary_bandit import *
 from .ordinary_mnl_bandit import *
+from .ordinary_mnl_bandit_utils import *
 from .linear_bandit import *
 from .thres_bandit import *
 from .contextual_bandit import *
